:art: refactor: Open API로 데이터 수집하는 기능, 알뜰가게 정보 제공 기능 분리
