+++ src/main/java/zzandori/zzanmoa/savingplace/dto/StoreInfoDTO.java
@@ -1,22 +1,25 @@
-package zzandori.zzanmoa.thriftstore.dto;
+package zzandori.zzanmoa.savingplace.dto;
 
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+
+import java.util.List;
 import lombok.ToString;
 
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
+@Builder
 @ToString
-public class StoreWithItemDTO {
-
+public class StoreInfoDTO {
     private String storeId;
     private String storeName;
-    private String category;
-    private String itemId;
     private String phoneNumber;
     private String address;
+    private List<ItemInfoDTO> items;
+
 }
