+++ src/main/java/zzandori/zzanmoa/savingplace/repository/SavingItemRepository.java
@@ -1,10 +1,10 @@
-package zzandori.zzanmoa.thriftstore.repository;
+package zzandori.zzanmoa.savingplace.repository;
 
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
-import zzandori.zzanmoa.thriftstore.dto.CategoryPriceDTO;
-import zzandori.zzanmoa.thriftstore.entity.SavingItem;
+import zzandori.zzanmoa.savingplace.dto.CategoryPriceDTO;
+import zzandori.zzanmoa.savingplace.entity.SavingItem;
 
 public interface SavingItemRepository extends JpaRepository<SavingItem, Long> {
