+++ src/main/java/zzandori/zzanmoa/savingplace/repository/SavingStoreRepository.java
@@ -1,9 +1,9 @@
-package zzandori.zzanmoa.thriftstore.repository;
+package zzandori.zzanmoa.savingplace.repository;
 
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
-import zzandori.zzanmoa.thriftstore.entity.SavingStore;
+import zzandori.zzanmoa.savingplace.entity.SavingStore;
 
 public interface SavingStoreRepository extends JpaRepository<SavingStore, Long> {
