+++ src/main/java/zzandori/zzanmoa/thriftstore/controller/ThriftStoreController.java
@@ -1,14 +1,11 @@
 package zzandori.zzanmoa.thriftstore.controller;
 
 import java.io.IOException;
-import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import zzandori.zzanmoa.thriftstore.dto.CategoryPriceDTO;
-import zzandori.zzanmoa.thriftstore.dto.StoreInfoDTO;
 import zzandori.zzanmoa.thriftstore.service.ThriftStoreService;
 
 @RestController
@@ -17,17 +14,6 @@ import zzandori.zzanmoa.thriftstore.service.ThriftStoreService;
 public class ThriftStoreController {
 
     private final ThriftStoreService thriftStoreService;
-
-
-    @GetMapping("/get/category")
-    public List<CategoryPriceDTO> getCategoryPrice(){
-        return thriftStoreService.getCategoryPrice();
-    }
-
-    @GetMapping("get/store")
-    public List<StoreInfoDTO> getStores(){
-        return thriftStoreService.getAllStoresWithItems();
-    }
 
 
     @GetMapping("/connectAPI/{startIndex}/{endIndex}")
